Tidying up and testing
