+++ Engine/Managers/SystemManager.cs
@@ -38,8 +38,8 @@ namespace AdventureGame.Engine
             AddSystem(new DamageSystem());
             AddSystem(new HealthSystem());
             AddSystem(new MapCollisionSystem());
-            AddSystem(new CollisionSystem3());
-            AddSystem(new CollisionResponseSystem3());
+            AddSystem(new CollisionSystem());
+            AddSystem(new CollisionResponseSystem());
             AddSystem(new TriggerSystem());
             AddSystem(new SpriteSystem());
             AddSystem(new AnimationSystem());
