+++ Engine/Systems/CollisionSystem.cs
@@ -80,6 +80,92 @@ namespace AdventureGame.Engine
             TestingOutputSets();
         }
 
+        public void AddCollisionHandler(Entity entity)
+        {
+            CollisionHandlerComponent handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
+
+            // Check if the entity isn't handling any collisions currently
+            if (handlerComponent == null)
+                entity.AddComponent(new CollisionHandlerComponent(), true);
+        }
+
+        public void AddToCollisionStarted(Entity entity, Entity otherEntity)
+        {
+            CollisionHandlerComponent handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
+
+            // Add a collision handler if the entity isn't handling any collisions currently
+            if (handlerComponent == null)
+            {
+                AddCollisionHandler(entity);
+                handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
+            }
+
+            // Add the entity to the collision started sets
+            _collisionStarted.Add(entity);
+            handlerComponent.CollidedEntities.Add(otherEntity);
+        }
+
+        public void RemoveFromCollisionStarted(Entity entity, Entity otherEntity)
+        {
+            CollisionHandlerComponent handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
+            if (handlerComponent == null)
+                return;
+
+            // Remove the collided entity from the component set
+            handlerComponent.CollidedEntities.Remove(otherEntity);
+
+            // Check if the entity has any more collisions to handle
+            if (handlerComponent.CollidedEntities.Count == 0)
+                _collisionStarted.Remove(entity);
+        }
+
+        public void AddToCollisionEnded(Entity entity, Entity otherEntity)
+        {
+            CollisionHandlerComponent handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
+            if (handlerComponent == null)
+                return;
+
+            // Add the entities to the collision ended sets
+            _collisionEnded.Add(entity);
+            handlerComponent.CollidedEntitiesEnded.Add(otherEntity);
+        }
+
+        public void RemoveFromCollisionEnded(Entity entity, Entity otherEntity)
+        {
+            CollisionHandlerComponent handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
+            if (handlerComponent == null)
+                return;
+
+            // Remove the collided entity from the component set
+            handlerComponent.CollidedEntitiesEnded.Remove(otherEntity);
+
+            // Check if the entity has any more ended collisions to handle
+            if (handlerComponent.CollidedEntitiesEnded.Count == 0)
+            {
+                _collisionEnded.Remove(entity);
+
+                if (handlerComponent.CollidedEntities.Count == 0)
+                    entity.RemoveComponent<CollisionHandlerComponent>();
+            }
+        }
+
+        public override void Update(GameTime gameTime, Scene scene)
+        {
+            foreach (Entity e in entityList)
+            {
+                TransformComponent transformComponent = e.GetComponent<TransformComponent>();
+
+                // Update the bounding box of all moving entities
+                if (transformComponent.HasMoved())
+                {
+                    ColliderComponent colliderComponent = e.GetComponent<ColliderComponent>();
+                    colliderComponent.GetBoundingBox(transformComponent.position);
+                }
+
+                // Broad-phase here too??
+            }
+        }
+
         public override void UpdateEntity(GameTime gameTime, Scene scene, Entity entity)
         {
             ColliderComponent colliderComponent = entity.GetComponent<ColliderComponent>();
@@ -92,11 +178,11 @@ namespace AdventureGame.Engine
                 colliderComponent.Broadphase = Rectangle.Empty;
 
             // Only check moving entities
-            if (!transformComponent.HasMoved() && !_collisionEnded.Contains(entity))
+            if (!transformComponent.HasMoved())// && !_collisionEnded.Contains(entity))
                 return;
 
             // Update the bounding box
-            colliderComponent.GetBoundingBox(transformComponent.position);
+            //colliderComponent.GetBoundingBox(transformComponent.position);
 
             // Testing: Broad-phasing
             Rectangle broadphaseBox;
@@ -111,28 +197,44 @@ namespace AdventureGame.Engine
                 if (entity != otherEntity)
                 {
                     ColliderComponent otherColliderComponent = otherEntity.GetComponent<ColliderComponent>();
+                    CollisionHandlerComponent otherHandlerComponent = otherEntity.GetComponent<CollisionHandlerComponent>();
+                    //PhysicsComponent otherPhysicsComponent = otherEntity.GetComponent<PhysicsComponent>(); Broadphase?
+                    TransformComponent otherTransformComponent = otherEntity.GetComponent<TransformComponent>();
 
                     // Check if the entities have collided
                     if (colliderComponent.Box.Intersects(otherColliderComponent.Box))
                     {
-                        //Console.WriteLine($"\nEntity {entity.Id} intersects with {otherEntity.Id}");
+                        Console.WriteLine($"\nEntity {entity.Id} intersects with {otherEntity.Id}");
 
                         // Check if the entity isn't handling any collisions currently
-                        if (handlerComponent == null)
+                        AddCollisionHandler(entity);
+                        handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
+                        /*if (handlerComponent == null)
                         {
                             entity.AddComponent(new CollisionHandlerComponent(), true);
                             handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
-                        }
+                        }*/
 
                         // Check if the entities are already colliding
                         if (!(_collisionStarted.Contains(entity)
                             && handlerComponent.CollidedEntities.Contains(otherEntity)))
                         {
-                            Console.WriteLine($"\nStart collision: {entity.Id} & {otherEntity.Id}");
+                            Console.WriteLine($"Start collision: {entity.Id} & {otherEntity.Id}");
 
                             // Add the entity to the collision started sets
-                            _collisionStarted.Add(entity);
-                            handlerComponent.CollidedEntities.Add(otherEntity);
+                            AddToCollisionStarted(entity, otherEntity);
+
+                            // Add the other entity if it is also moving
+                            //if (otherTransformComponent.HasMoved())
+                            //    AddToCollisionStarted(otherEntity, entity);
+
+                            // Check if this collision ended last frame
+                            if (_collisionEnded.Contains(entity)
+                                && handlerComponent.CollidedEntitiesEnded.Contains(otherEntity))
+                            {
+                                RemoveFromCollisionEnded(entity, otherEntity);
+                                //RemoveFromCollisionEnded(otherEntity, entity);
+                            }
 
                             // Testing: change component outline colour
                             colliderComponent.color = Color.Orange;
@@ -146,39 +248,33 @@ namespace AdventureGame.Engine
                     else if (_collisionStarted.Contains(entity) && handlerComponent != null
                         && handlerComponent.CollidedEntities.Contains(otherEntity))
                     {
-                        Console.WriteLine($"End collision: {entity.Id} & {otherEntity.Id}");
+                        Console.WriteLine($"\nEnd collision: {entity.Id} & {otherEntity.Id}");
 
-                        // Remove the collided entity from the component set
-                        handlerComponent.CollidedEntities.Remove(otherEntity);
+                        RemoveFromCollisionStarted(entity, otherEntity);
+                        AddToCollisionEnded(entity, otherEntity);
 
-                        // Check if the entity has any more collisions to handle
-                        if (handlerComponent.CollidedEntities.Count == 0)
-                            _collisionStarted.Remove(entity);
-
-                        // Add the entities to the collision ended sets
-                        _collisionEnded.Add(entity);
-                        handlerComponent.CollidedEntitiesEnded.Add(otherEntity);
+                        // Check if the other entity was handling collisions
+                        //if (otherHandlerComponent != null)
+                        //{
+                        //    RemoveFromCollisionStarted(otherEntity, entity);
+                        //    AddToCollisionEnded(otherEntity, entity);
+                        //}
 
                         TestingOutputSets(); // Testing
                     }
 
+                    // TO DO - should it also check collisionStarted / collidedEntities??
                     // Check if the entites are exiting a previous collision
                     else if (_collisionEnded.Contains(entity) && handlerComponent != null
                         && handlerComponent.CollidedEntitiesEnded.Contains(otherEntity))
                     {
-                        Console.WriteLine($"Exiting collision: {entity.Id} & {otherEntity.Id}");
+                        Console.WriteLine($"\nExiting collision: {entity.Id} & {otherEntity.Id}");
 
-                        // Remove the collided entity from the component set
-                        handlerComponent.CollidedEntitiesEnded.Remove(otherEntity);
+                        RemoveFromCollisionEnded(entity, otherEntity);
 
-                        // Check if the entity has any more ended collisions to handle
-                        if (handlerComponent.CollidedEntitiesEnded.Count == 0)
-                        {
-                            _collisionEnded.Remove(entity);
-
-                            if (handlerComponent.CollidedEntities.Count == 0)
-                                entity.RemoveComponent<CollisionHandlerComponent>();
-                        }
+                        // Check if the other entity was handling collisions
+                        //if (otherHandlerComponent != null)
+                        //    RemoveFromCollisionEnded(otherEntity, entity);
 
                         // Testing: change component outline colour
                         colliderComponent.color = Color.Yellow;
@@ -212,7 +308,7 @@ namespace AdventureGame.Engine
         public void TestingOutputSets()
         {
             // Testing: output collided entities set
-            Console.WriteLine("\nCollision started set: ");
+            Console.WriteLine("Collision started set: ");
             foreach (Entity entity in _collisionStarted)
             {
                 CollisionHandlerComponent handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
@@ -228,7 +324,7 @@ namespace AdventureGame.Engine
             }
 
             // Testing: output collided entities ended set
-            Console.WriteLine("\nCollision ended set: ");
+            Console.WriteLine("Collision ended set: ");
             foreach (Entity entity in _collisionEnded)
             {
                 CollisionHandlerComponent handlerComponent = entity.GetComponent<CollisionHandlerComponent>();
